reorganize the way functions were being called
I changed results.asList() so it called it only once instead of twice. I also made it so if the number of queries go above 10 or below 0, it defaults back to 10. Because of this change, I edited the getLimit function so it only returns -1 when there is something other than an int passed (since the parent function can now handle negative and out of range values). I also created a variable name for 10.